Move attributes to style module
